Clear timeout handlers.
It is known that setTimeout(0[, callback]) must remove listeners to avoid
memory leaks, as warnings and leaks were encountered.

It's clear that setTimeout(0, callback) means "remove listener". But should
other timeout listeners remain in place to "turn on" again with a non-zero
timeout? That's what will happen with this implementation.

Does setTimeout(0) mean "remove all listeners" or just "temporarily disable
them"? This implementation does the former.